git add Makefile linker.ld src/isr_stubs.s src/task.h src/task.c src/timer.c
git add Makefile linker.ld src/isr_stubs.s src/task.h src/task.c src/timer.c
git commit -m "Core: 32-bit tick scheduler, GAS ISR stubs and gcc-only build"
git push

+++ src/task.h
@@ -12,8 +12,8 @@ void tasks_init(void);
 
 /*
  * Вызывается по каждому тиканью таймера (IRQ0).
- * tick — глобальный счётчик тиков.
+ * tick — глобальный 32-битный счётчик тиков.
  */
-void scheduler_on_tick(uint64_t tick);
+void scheduler_on_tick(uint32_t tick);
 
 #endif // BOLGENOS_TASK_H
